SparkBot_AI/requirements.txt
flask
python-dotenv
google-generativeai
werkzeug
